refactor: merge existing LDV and HDV implementations into a common codebase (#318)
