+++ prereise/gather/demanddata/transportation_electrification/tests/hdv_test_data.csv
@@ -1,0 +1,50 @@
+,Vehicle Number,Trip Number,Home base end (1/0),Destination from,Destination to,Trip Distance,Trip Start,Trip End,Dwell Time,Trip Time,Speed,Total Vehicle Trips,Total Vehicle Miles
+0,1.0,1.0,2.0,1.0,2.0,4.986796239,7.0,7.333333333,0.916666667,0.333333333,14.96038873,5.0,39.55650414
+1,1.0,2.0,2.0,2.0,2.0,5.339196297,8.25,10.16666667,0.3,1.91666667,2.785667628,5.0,39.55650414
+2,1.0,3.0,2.0,2.0,2.0,13.17360767,10.46666667,10.83333333,0.25,0.36666666,35.92802157,5.0,39.55650414
+3,1.0,4.0,2.0,2.0,2.0,8.895112039,11.08333333,13.75,0.41666667,2.66666667,3.33566701,5.0,39.55650414
+4,1.0,5.0,1.0,2.0,1.0,7.161791891,14.16666667,14.75,16.25,0.58333333,12.2773576,5.0,39.55650414
+5,3.0,1.0,2.0,1.0,2.0,3.218623172,6.466666667,7.0,0.25,0.533333333,6.034918451,5.0,37.90529228
+6,3.0,2.0,2.0,2.0,2.0,10.15727088,7.25,8.0,0.916666667,0.75,13.54302784,5.0,37.90529228
+7,3.0,3.0,2.0,2.0,2.0,15.62816085,8.916666667,10.83333333,0.5,1.916666663,8.153823068,5.0,37.90529228
+8,3.0,4.0,2.0,2.0,2.0,8.494012972,11.33333333,11.75,1.0,0.41666667,20.38563097,5.0,37.90529228
+9,3.0,5.0,1.0,2.0,1.0,0.407224407,12.75,13.0,17.46666667,0.25,1.628897628,5.0,37.90529228
+10,6.0,1.0,2.0,1.0,2.0,15.47773884,9.266666667,9.8,0.25,0.533333333,29.02076034,7.0,33.35874777
+11,6.0,2.0,2.0,2.0,3.0,0.18516606,10.05,10.25,0.4,0.2,0.9258303,7.0,33.35874777
+12,6.0,3.0,2.0,3.0,3.0,0.353662028,10.65,10.83333333,0.43333334,0.18333333,1.929065642,7.0,33.35874777
+13,6.0,4.0,2.0,3.0,3.0,1.087056837,11.26666667,11.6,0.56666667,0.33333333,3.261170544,7.0,33.35874777
+14,6.0,5.0,2.0,3.0,3.0,1.025934145,12.16666667,12.6,0.2,0.43333333,2.367540353,7.0,33.35874777
+15,6.0,6.0,2.0,3.0,3.0,0.907816224,12.8,13.2,0.48333333,0.4,2.26954056,7.0,33.35874777
+16,6.0,7.0,1.0,3.0,1.0,14.32137364,13.68333333,14.28333333,18.98333334,0.6,23.86895607,7.0,33.35874777
+17,37.0,1.0,1.0,1.0,1.0,35.83017697,11.75,13.0,22.75,1.25,28.66414158,1.0,35.83017697
+18,39.0,1.0,2.0,1.0,7.0,9.232786345,5.0,5.5,8.0,0.5,18.46557269,2.0,32.76195438
+19,39.0,2.0,1.0,7.0,1.0,23.52916803,13.5,14.5,14.5,1.0,23.52916803,2.0,32.76195438
+20,44.0,1.0,2.0,1.0,2.0,9.157182922,8.083333333,8.25,0.666666667,0.166666667,54.94309742,12.0,42.20693609
+21,44.0,2.0,2.0,2.0,2.0,3.054817837,8.916666667,9.333333333,0.166666667,0.416666666,7.331562821,12.0,42.20693609
+22,44.0,3.0,2.0,2.0,2.0,0.716574493,9.5,9.583333333,0.166666667,0.083333333,8.59889395,12.0,42.20693609
+23,44.0,4.0,2.0,2.0,2.0,0.723409635,9.75,9.916666667,0.333333333,0.166666667,4.340457801,12.0,42.20693609
+24,44.0,5.0,2.0,2.0,2.0,3.388731346,10.25,11.08333333,0.5,0.83333333,4.066477631,12.0,42.20693609
+25,44.0,6.0,2.0,2.0,2.0,2.35438895,11.58333333,12.08333333,0.33333334,0.5,4.7087779,12.0,42.20693609
+26,44.0,7.0,2.0,2.0,2.0,0.719757468,12.41666667,12.66666667,0.25,0.25,2.879029872,12.0,42.20693609
+27,44.0,8.0,2.0,2.0,2.0,0.624738745,12.91666667,13.66666667,0.33333333,0.75,0.832984993,12.0,42.20693609
+28,44.0,9.0,2.0,2.0,2.0,8.738050447,14.0,14.75,1.0,0.75,11.65073393,12.0,42.20693609
+29,44.0,10.0,2.0,2.0,2.0,0.055260226,15.75,16.25,0.25,0.5,0.110520452,12.0,42.20693609
+30,44.0,11.0,2.0,2.0,2.0,2.201842476,16.5,16.66666667,0.75,0.16666667,13.21105459,12.0,42.20693609
+31,44.0,12.0,1.0,2.0,1.0,10.47218154,17.41666667,17.98333333,14.1,0.56666666,18.48032058,12.0,42.20693609
+32,45.0,1.0,2.0,1.0,2.0,16.00014886,9.0,9.583333333,0.416666667,0.583333333,27.42882663,7.0,86.47557714
+33,45.0,2.0,1.0,2.0,3.0,15.64956143,10.0,10.66666667,1.33333333,0.66666667,23.47434203,7.0,86.47557714
+34,45.0,3.0,2.0,3.0,6.0,3.169170578,12.0,12.16666667,0.75,0.16666667,19.01502309,7.0,86.47557714
+35,45.0,4.0,2.0,6.0,2.0,22.52937084,12.91666667,13.66666667,1.0,0.75,30.03916112,7.0,86.47557714
+36,45.0,5.0,1.0,2.0,3.0,18.84876732,14.66666667,15.5,1.25,0.83333333,22.61852087,7.0,86.47557714
+37,45.0,6.0,2.0,3.0,2.0,5.297067936,16.75,17.08333333,0.25,0.33333333,15.89120397,7.0,86.47557714
+38,45.0,7.0,1.0,2.0,1.0,4.981490175,17.33333333,17.66666667,15.33333333,0.33333334,14.94447023,7.0,86.47557714
+39,47.0,1.0,2.0,1.0,2.0,10.72104885,5.0,5.5,0.666666667,0.5,21.4420977,5.0,67.56448333
+40,47.0,2.0,1.0,2.0,3.0,10.40785863,6.166666667,6.75,1.75,0.583333333,17.84204338,5.0,67.56448333
+41,47.0,3.0,2.0,3.0,2.0,17.76320637,8.5,9.083333333,1.083333337,0.583333333,30.45121094,5.0,67.56448333
+42,47.0,4.0,2.0,2.0,2.0,12.33972913,10.16666667,10.75,1.5,0.58333333,21.15382149,5.0,67.56448333
+43,47.0,5.0,1.0,2.0,1.0,16.33264035,12.25,13.0,16.0,0.75,21.7768538,5.0,67.56448333
+44,48.0,1.0,2.0,1.0,2.0,5.476941163,4.5,5.25,1.0,0.75,7.302588217,5.0,53.83477584
+45,48.0,2.0,1.0,2.0,3.0,5.794782464,6.25,6.75,1.25,0.5,11.58956493,5.0,53.83477584
+46,48.0,3.0,2.0,3.0,2.0,5.339828215,8.0,8.333333333,0.916666667,0.333333333,16.01948466,5.0,53.83477584
+47,48.0,4.0,2.0,2.0,2.0,17.71778974,9.25,10.25,0.25,1.0,17.71778974,5.0,53.83477584
+48,48.0,5.0,1.0,2.0,1.0,19.50543426,10.5,11.75,16.75,1.25,15.60434741,5.0,53.83477584
