+++ prereise/gather/demanddata/transportation_electrification/tests/mdv_test_data.csv
@@ -1,0 +1,50 @@
+,Vehicle Number,Trip Number,Home base end (1/0),Destination from,Destination to,Trip Distance,Trip Start,Trip End,Dwell Time,Trip Time,Speed,Total Vehicle Trips,Total Vehicle Miles
+0,174.0,1.0,2.0,1.0,4.0,2.951962304,7.75,8.666666667,2.350000003,0.916666667,3.220322512,2.0,5.903372153
+1,174.0,2.0,1.0,4.0,1.0,2.951409849,11.01666667,11.5,20.25,0.483333333,6.106365247,2.0,5.903372153
+2,201.0,1.0,2.0,1.0,2.0,4.731915611,9.25,9.666666667,8.833333333,0.416666667,11.35659746,2.0,9.457050469
+3,201.0,2.0,1.0,2.0,1.0,4.725134858,18.5,19.0,14.25,0.5,9.450269716,2.0,9.457050469
+4,372.0,1.0,2.0,1.0,2.0,1.00592117,13.5,13.75,2.75,0.25,4.02368468,3.0,4.394385517
+5,372.0,2.0,2.0,2.0,2.0,2.108335902,16.5,16.75,4.41666667,0.25,8.433343608,3.0,4.394385517
+6,372.0,3.0,1.0,2.0,1.0,1.280128445,21.16666667,21.33333333,16.16666667,0.166666667,7.680770977,3.0,4.394385517
+7,373.0,1.0,2.0,1.0,2.0,1.886776554,7.75,8.0,3.33333333,0.25,7.547106216,2.0,3.88381754
+8,373.0,2.0,1.0,2.0,1.0,1.997040986,11.33333333,11.5,20.25,0.166666667,11.98224568,2.0,3.88381754
+9,707.0,1.0,2.0,1.0,8.0,0.326182794,14.71666667,14.75,0.2,0.03333333,9.785484799,6.0,6.403946313
+10,707.0,2.0,2.0,8.0,8.0,0.574355639,14.95,14.96666667,0.01666666,0.01666667,34.46133145,6.0,6.403946313
+11,707.0,3.0,2.0,8.0,8.0,0.5610885,14.98333333,15.01666667,0.13333333,0.03333334,16.83265163,6.0,6.403946313
+12,707.0,4.0,2.0,8.0,8.0,1.683588643,15.15,15.28333333,0.1,0.13333333,12.62691514,6.0,6.403946313
+13,707.0,5.0,2.0,8.0,8.0,1.193255391,15.38333333,15.48333333,0.03333334,0.1,11.93255391,6.0,6.403946313
+14,707.0,6.0,1.0,8.0,1.0,2.065475346,15.51666667,15.61666667,23.1,0.1,20.65475346,6.0,6.403946313
+15,708.0,1.0,2.0,1.0,8.0,1.009143759,13.83333333,13.91666667,0.13333333,0.08333334,12.10972414,6.0,9.227443902
+16,708.0,2.0,2.0,8.0,5.0,0.453613052,14.05,14.58333333,0.1,0.53333333,0.850524478,6.0,9.227443902
+17,708.0,3.0,2.0,5.0,8.0,0.967425404,14.68333333,14.83333333,0.18333334,0.15,6.449502693,6.0,9.227443902
+18,708.0,4.0,2.0,8.0,8.0,1.829310337,15.01666667,15.18333333,0.01666667,0.16666666,10.97586246,6.0,9.227443902
+19,708.0,5.0,2.0,8.0,8.0,2.412512776,15.2,15.38333333,0.06666667,0.18333333,13.15916084,6.0,9.227443902
+20,708.0,6.0,1.0,8.0,1.0,2.555438574,15.45,15.61666667,22.21666666,0.16666667,15.33263114,6.0,9.227443902
+21,710.0,1.0,2.0,1.0,8.0,2.087751178,4.58333333,4.733333333,10.35,0.150000003,13.91834091,2.0,4.476653447
+22,710.0,2.0,1.0,8.0,1.0,2.388902269,15.08333333,15.16666667,13.41666666,0.08333334,28.66682493,2.0,4.476653447
+23,731.0,1.0,2.0,1.0,4.0,1.813810919,7.0,9.0,10.0,2.0,0.90690546,2.0,3.494545172
+24,731.0,2.0,1.0,4.0,1.0,1.680734253,19.0,21.0,10.0,2.0,0.840367127,2.0,3.494545172
+25,733.0,1.0,2.0,1.0,8.0,1.813810919,9.0,9.5,7.0,0.5,3.627621838,2.0,3.494545172
+26,733.0,2.0,1.0,8.0,1.0,1.680734253,16.5,17.0,16.0,0.5,3.361468506,2.0,3.494545172
+27,734.0,1.0,2.0,1.0,4.0,1.830973255,7.916666667,8.083333333,0.833333334,0.166666666,10.98583957,2.0,5.470067658
+28,734.0,2.0,1.0,4.0,1.0,3.639094403,8.916666667,9.083333333,22.83333333,0.166666666,21.83456651,2.0,5.470067658
+29,737.0,1.0,2.0,1.0,8.0,3.420729653,13.25,13.5,1.75,0.25,13.68291861,3.0,11.11683692
+30,737.0,2.0,2.0,8.0,8.0,2.137688802,15.25,15.33333333,0.66666667,0.08333333,25.65226665,3.0,11.11683692
+31,737.0,3.0,1.0,8.0,1.0,5.558418462,16.0,16.25,21.0,0.25,22.23367385,3.0,11.11683692
+32,747.0,1.0,1.0,1.0,1.0,5.828878996,9.25,9.5,23.75,0.25,23.31551598,1.0,5.828878996
+33,748.0,1.0,1.0,1.0,1.0,5.828878996,18.0,18.5,23.5,0.5,11.65775799,1.0,5.828878996
+34,752.0,1.0,2.0,1.0,7.0,2.61261866,7.5,7.666666667,8.333333333,0.166666667,15.67571193,2.0,5.188354235
+35,752.0,2.0,1.0,7.0,1.0,2.575735575,16.0,16.16666667,15.33333333,0.16666667,15.45441314,2.0,5.188354235
+36,759.0,1.0,2.0,1.0,10.0,1.719509247,7.0,7.05,0.15,0.05,34.39018494,4.0,7.907749186
+37,759.0,2.0,2.0,10.0,10.0,1.26557955,7.2,7.333333333,0.016666667,0.133333333,9.491846649,4.0,7.907749186
+38,759.0,3.0,2.0,10.0,10.0,1.218749131,7.35,7.416666667,0.133333333,0.066666667,18.28123687,4.0,7.907749186
+39,759.0,4.0,1.0,10.0,1.0,3.703911258,7.55,7.75,23.25,0.2,18.51955629,4.0,7.907749186
+40,760.0,1.0,2.0,1.0,10.0,2.660773721,6.833333333,7.083333333,0.016666667,0.25,10.64309488,2.0,5.455996612
+41,760.0,2.0,1.0,10.0,1.0,2.795222891,7.1,7.35,23.48333333,0.25,11.18089156,2.0,5.455996612
+42,761.0,1.0,2.0,1.0,10.0,0.314803389,7.0,7.083333333,0.016666667,0.083333333,3.777640683,5.0,7.46392925
+43,761.0,2.0,2.0,10.0,10.0,1.259954291,7.1,7.183333333,0.016666667,0.083333333,15.11945155,5.0,7.46392925
+44,761.0,3.0,2.0,10.0,10.0,2.170176764,7.2,7.333333333,0.016666667,0.133333333,16.27632577,5.0,7.46392925
+45,761.0,4.0,2.0,10.0,10.0,1.905821571,7.35,7.433333333,0.016666667,0.083333333,22.86985894,5.0,7.46392925
+46,761.0,5.0,1.0,10.0,1.0,1.813173235,7.45,7.616666667,23.38333333,0.166666667,10.87903939,5.0,7.46392925
+47,769.0,1.0,2.0,1.0,4.0,0.611258646,13.5,13.66666667,0.96666666,0.16666667,3.667551803,2.0,1.193933971
+48,769.0,2.0,1.0,4.0,1.0,0.582675325,14.63333333,14.8,22.7,0.16666667,3.49605188,2.0,1.193933971
