+++ prereise/gather/demanddata/transportation_electrification/tests/test_smart_charging.py
@@ -1,171 +1,22 @@
-import inspect
 import os
 
 import numpy as np
-import pytest
 from scipy.io import loadmat
 
-import prereise
-import prereise.gather.demanddata.transportation_electrification.immediate_charging_HDV as immediate_charging_HDV
 from prereise.gather.demanddata.transportation_electrification import (
+    const,
     data_helper,
     smart_charging,
     smart_charging_HDV,
 )
 from prereise.gather.demanddata.transportation_electrification.data_helper import (
     generate_daily_weighting,
-    load_urbanized_scaling_factor,
 )
-from prereise.gather.demanddata.transportation_electrification.immediate import (
-    immediate_charging,
-)
-
-
-def test_immediate_charging_region1():
-    immediate_charging(
-        census_region=1,
-        model_year=2017,
-        veh_range=100,
-        kwhmi=0.242,
-        power=6.6,
-        location_strategy=2,
-        veh_type="LDV",
-        filepath=os.path.join(
-            os.path.dirname(inspect.getsourcefile(prereise)),
-            "gather",
-            "demanddata",
-            "transportation_electrification",
-            "data",
-            "nhts_census_updated_dwell",
-        ),
-    )
-
-
-def test_immediate_charging_mdv():
-
-    result = immediate_charging_HDV.immediate_charging(
-        model_year=2050,
-        veh_range=200,
-        power=80,
-        location_strategy=1,
-        veh_type="MDV",
-        filepath=os.path.join(
-            os.path.dirname(inspect.getsourcefile(prereise)),
-            "gather",
-            "demanddata",
-            "transportation_electrification",
-            "data",
-            "fdata_v10st.mat",
-        ),
-        trip_strategy=1,
-    )
-    bev_vmt = load_urbanized_scaling_factor(
-        model_year=2050,
-        veh_type="MDV",
-        veh_range=200,
-        urbanized_area="Antioch",
-        state="CA",
-        filepath=os.path.join(
-            os.path.dirname(inspect.getsourcefile(prereise)),
-            "gather",
-            "demanddata",
-            "transportation_electrification",
-            "data",
-            "regional_scaling_factors",
-            "Regional_scaling_factors_UA_",
-        ),
-    )
-    final_result = immediate_charging_HDV.adjust_bev(
-        model_year=2050,
-        veh_type="MDV",
-        veh_range=200,
-        model_year_profile=result,
-        bev_vmt=bev_vmt,
-        charging_efficiency=0.95,
-    )
-
-    correct_cumsum = np.array(
-        [
-            5.61689215500168,
-            10942.68623,
-            21951.59332,
-            32980.21489,
-            43904.32619,
-            54934.05497,
-            65925.06389,
-            76881.00051,
-        ],
-    )
-
-    np.testing.assert_allclose(final_result.cumsum()[::1095], correct_cumsum)
-
-
-def test_immediate_charging_hdv():
-
-    result = immediate_charging_HDV.immediate_charging(
-        model_year=2050,
-        veh_range=200,
-        power=80,
-        location_strategy=1,
-        veh_type="HDV",
-        filepath=os.path.join(
-            os.path.dirname(inspect.getsourcefile(prereise)),
-            "gather",
-            "demanddata",
-            "transportation_electrification",
-            "data",
-            "fdata_v10st.mat",
-        ),
-        trip_strategy=1,
-    )
-    bev_vmt = load_urbanized_scaling_factor(
-        model_year=2050,
-        veh_type="HDV",
-        veh_range=200,
-        urbanized_area="Antioch",
-        state="CA",
-        filepath=os.path.join(
-            os.path.dirname(inspect.getsourcefile(prereise)),
-            "gather",
-            "demanddata",
-            "transportation_electrification",
-            "data",
-            "regional_scaling_factors",
-            "Regional_scaling_factors_UA_",
-        ),
-    )
-    final_result = immediate_charging_HDV.adjust_bev(
-        model_year=2050,
-        veh_type="HDV",
-        veh_range=200,
-        model_year_profile=result,
-        bev_vmt=bev_vmt,
-        charging_efficiency=0.95,
-    )
-
-    correct_cumsum = np.array(
-        [
-            10.8255287099344,
-            10458.16992,
-            20964.88916,
-            31443.09284,
-            41900.81189,
-            52407.05727,
-            62866.27995,
-            73352.53979,
-        ],
-    )
-
-    np.testing.assert_allclose(final_result.cumsum()[::1095], correct_cumsum)
 
 
 def test_smart_charging():
     data_dir = os.path.join(
-        os.path.dirname(inspect.getsourcefile(prereise)),
-        "gather",
-        "demanddata",
-        "transportation_electrification",
-        "data",
+        const.data_folder_path,
         "CAISO_sample_load_2019.mat",
     )
     load_demand = loadmat(data_dir)["load_demand"].flatten()
@@ -181,12 +32,8 @@ def test_smart_charging():
         location_strategy=2,
         veh_type="LDV",
         filepath=os.path.join(
-            os.path.dirname(inspect.getsourcefile(prereise)),
-            "gather",
-            "demanddata",
-            "transportation_electrification",
-            "tests",
-            "test_census_data.csv",
+            const.test_folder_path,
+            "ldv_test_data.csv",
         ),
         daily_values=daily_values,
         load_demand=load_demand,
@@ -209,14 +56,9 @@ def test_smart_charging():
     np.testing.assert_allclose(result.cumsum()[::1095], correct_cumsum)
 
 
-@pytest.mark.skip(reason="currently too slow")
 def test_smart_charging_hdv():
     data_dir = os.path.join(
-        os.path.dirname(inspect.getsourcefile(prereise)),
-        "gather",
-        "demanddata",
-        "transportation_electrification",
-        "data",
+        const.data_folder_path,
         "CAISO_sample_load_2019.mat",
     )
     load_demand = loadmat(data_dir)["load_demand"].flatten()
@@ -228,11 +70,7 @@ def test_smart_charging_hdv():
         urbanized_area="Antioch",
         state="CA",
         filepath=os.path.join(
-            os.path.dirname(inspect.getsourcefile(prereise)),
-            "gather",
-            "demanddata",
-            "transportation_electrification",
-            "data",
+            const.data_folder_path,
             "regional_scaling_factors",
             "Regional_scaling_factors_UA_",
         ),
@@ -244,12 +82,8 @@ def test_smart_charging_hdv():
         location_strategy=1,
         veh_type="HDV",
         filepath=os.path.join(
-            os.path.dirname(inspect.getsourcefile(prereise)),
-            "gather",
-            "demanddata",
-            "transportation_electrification",
-            "data",
-            "fdata_v10st.mat",
+            const.test_folder_path,
+            "hdv_test_data.csv",
         ),
         initial_load=load_demand,
         bev_vmt=bev_vmt,
@@ -258,28 +92,23 @@ def test_smart_charging_hdv():
 
     correct_cumsum = np.array(
         [
-            1.22854177233283,
-            4729.417063,
-            9456.028814,
-            14087.49171,
-            18817.56654,
-            23521.75604,
-            28175.75066,
-            32904.52077,
+            0.0,
+            3856.38868062,
+            7712.77736124,
+            11485.33150533,
+            15341.72018595,
+            19197.26691589,
+            22970.66301066,
+            26827.05169128,
         ]
     )
 
     np.testing.assert_allclose(result.cumsum()[::1095], correct_cumsum)
 
 
-@pytest.mark.skip(reason="currently too slow")
 def test_smart_charging_mdv():
     data_dir = os.path.join(
-        os.path.dirname(inspect.getsourcefile(prereise)),
-        "gather",
-        "demanddata",
-        "transportation_electrification",
-        "data",
+        const.data_folder_path,
         "CAISO_sample_load_2019.mat",
     )
     load_demand = loadmat(data_dir)["load_demand"].flatten()
@@ -291,11 +120,7 @@ def test_smart_charging_mdv():
         urbanized_area="Antioch",
         state="CA",
         filepath=os.path.join(
-            os.path.dirname(inspect.getsourcefile(prereise)),
-            "gather",
-            "demanddata",
-            "transportation_electrification",
-            "data",
+            const.data_folder_path,
             "regional_scaling_factors",
             "Regional_scaling_factors_UA_",
         ),
@@ -307,12 +132,8 @@ def test_smart_charging_mdv():
         location_strategy=1,
         veh_type="MDV",
         filepath=os.path.join(
-            os.path.dirname(inspect.getsourcefile(prereise)),
-            "gather",
-            "demanddata",
-            "transportation_electrification",
-            "data",
-            "fdata_v10st.mat",
+            const.test_folder_path,
+            "mdv_test_data.csv",
         ),
         initial_load=load_demand,
         bev_vmt=bev_vmt,
@@ -321,14 +142,14 @@ def test_smart_charging_mdv():
 
     correct_cumsum = np.array(
         [
-            0.291225084060603,
-            4160.770734,
-            8320.031013,
-            12391.66618,
-            16553.00332,
-            20705.85663,
-            24784.48369,
-            28946.34624,
+            0.0,
+            4043.9900878,
+            8087.9801756,
+            12044.05743541,
+            16088.04752321,
+            20132.03761101,
+            24088.11487082,
+            28132.10495862,
         ]
     )
